+++ plugin-gradle/src/main/java/com/diffplug/gradle/spotless/SpotlessApply.java
@@ -20,38 +20,18 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.StandardCopyOption;
 
-import org.gradle.api.DefaultTask;
 import org.gradle.api.file.ConfigurableFileTree;
 import org.gradle.api.file.FileVisitDetails;
 import org.gradle.api.file.FileVisitor;
-import org.gradle.api.tasks.Internal;
 import org.gradle.api.tasks.TaskAction;
 
-public class SpotlessApply extends DefaultTask {
-	private SpotlessTask source;
-
-	/** Bidirectional link between Apply and Spotless allows check to know if Apply ran or not. */
-	void linkSource(SpotlessTask source) {
-		this.source = source;
-		source.applyTask = this;
-	}
-
-	private File spotlessOutDirectory;
-
-	@Internal
-	public File getSpotlessOutDirectory() {
-		return spotlessOutDirectory;
-	}
-
-	public void setSpotlessOutDirectory(File spotlessOutDirectory) {
-		this.spotlessOutDirectory = spotlessOutDirectory;
-	}
-
+public abstract class SpotlessApply extends SpotlessTaskService.ClientTask {
 	@TaskAction
 	public void performAction() {
-		ConfigurableFileTree files = getProject().fileTree(spotlessOutDirectory);
+		getTaskService().get().registerApplyAlreadyRan(this);
+		ConfigurableFileTree files = getConfigCacheWorkaround().fileTree().from(getSpotlessOutDirectory().get());
 		if (files.isEmpty()) {
-			getState().setDidWork(source.getDidWork());
+			getState().setDidWork(sourceDidWork());
 		} else {
 			files.visit(new FileVisitor() {
 				@Override
@@ -62,7 +42,7 @@ public class SpotlessApply extends DefaultTask {
 				@Override
 				public void visitFile(FileVisitDetails fileVisitDetails) {
 					String path = fileVisitDetails.getPath();
-					File originalSource = new File(getProject().getProjectDir(), path);
+					File originalSource = new File(getProjectDir().get().getAsFile(), path);
 					try {
 						getLogger().debug("Copying " + fileVisitDetails.getFile() + " to " + originalSource);
 						Files.copy(fileVisitDetails.getFile().toPath(), originalSource.toPath(), StandardCopyOption.REPLACE_EXISTING, StandardCopyOption.COPY_ATTRIBUTES);